Move endpoint tests to folders
