+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.IntegrationTests/IntegrationTests/EndpointTests/ChargeLinks/ChargeLinksCommandReceiverEndpointTests.cs
@@ -33,7 +33,7 @@ using Xunit;
 using Xunit.Abstractions;
 using Xunit.Categories;
 
-namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.EndpointTests
+namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.EndpointTests.ChargeLinks
 {
     [IntegrationTest]
     public class ChargeLinksCommandReceiverEndpointTests
