+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.IntegrationTests/IntegrationTests/EndpointTests/ChargeLinks/CreateDefaultChargeLinksReplierEndpointTests.cs
@@ -31,7 +31,7 @@ using Xunit;
 using Xunit.Abstractions;
 using Xunit.Categories;
 
-namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.EndpointTests
+namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.EndpointTests.ChargeLinks
 {
     [IntegrationTest]
     public class CreateDefaultChargeLinksReplierEndpointTests
